Use Rcpp automatic inclusion of _types header
We use the fact that Rcpp will now automatically include a specially named _types header in RcppExports.cpp to eliminate inst/include/RSQLite.h. It's contents are now broken up into three files in src: SqliteConnection.hpp, SqliteResult.hpp, and SqliteUtils.hpp
