+++ .travis.yml
@@ -13,7 +13,7 @@ before_install:
 install:
   - ./travis-tool.sh install_github rstats-db/DBI
   - ./travis-tool.sh install_github rstats-db/SQL
-  - ./travis-tool.sh install_r Rcpp
+  - ./travis-tool.sh install_github RcppCore/Rcpp
   - ./travis-tool.sh install_deps
   - R CMD INSTALL .
