--- inst/include/RSQLite.h
@@ -1,6 +1,0 @@
-/* 
-Rcpp needs this file to be included in RcppExports.cpp
-See: https://stackoverflow.com/questions/25494241/compileattributes-doesnt-copy-the-local-header-to-rcppexports-cpp
-*/
-
-#include "../../src/RSQLite.h"
