+++ src/RSQLite_types.hpp
@@ -1,0 +1,7 @@
+#ifndef __RSQLSITE_TYPES__
+#define __RSQLSITE_TYPES__
+
+#include "SqliteResult.hpp"
+#include "SqliteConnection.hpp"
+
+#endif // __RSQLSITE_TYPES__
