+++ src/RcppExports.cpp
@@ -1,7 +1,7 @@
 // This file was generated by Rcpp::compileAttributes
 // Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-#include "../inst/include/RSQLite.h"
+#include "RSQLite_types.hpp"
 #include <Rcpp.h>
 
 using namespace Rcpp;
